Detector_client/requirements.txt
numpy
tensorflow-gpu
imutils
pillow
hyperlpr